Refac: cleaned up core by removing steam-specific methods
